Isolate auto-lock functionality
